+++ src/Lakona.Game.Server/LakonaGameServerServiceCollectionExtensions.cs
@@ -1,6 +1,7 @@
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
+using Microsoft.Extensions.Logging;
 using Lakona.Game.Cluster;
 using Lakona.Game.Cluster.Rpc.Membership;
 using Lakona.Game.Server.Actors;
@@ -136,7 +137,8 @@ public static class LakonaGameServerServiceCollectionExtensions
                 participants,
                 provider.GetRequiredService<IClusterMembershipTransport>(),
                 membershipOptions,
-                provider);
+                provider,
+                provider.GetRequiredService<ILogger<ReplicatedClusterMembershipHostedService>>());
         });
         services.TryAddSingleton<IClusterMembership>(provider =>
             provider.GetRequiredService<ReplicatedClusterMembershipHostedService>());